Add files via upload
As Northeastern students who struggled with finding decently priced housing close enough to campus, we were curious to see what apartments were available in the area. The Northeastern community lacked a good database with convenient housing options. Through conversation with our peers, we realized this was a high demand prospect. As a result, we got together to compile a list of this database and analyzed it. We wanted to visualize where the most apartments were near campus, how highly priced they were and whether or not they were worth the ratings they were given. 
	Since we wanted to be able to compare the Northeastern data to another set, we chose to compare it to Harvard University data, since it was far away enough but still in the Higher Boston area. After analysis and visualization, we hope to see where the better housing options are: Northeastern or Harvard?